+++ Achievements/Conditions/NpcDropCondition.cs
@@ -27,7 +27,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <summary>
         /// NPC ID of the NPC that should drop the item(s)
         /// </summary>
-        private short _npcId;
+        private int _npcId;
 
 
         /// <summary>
@@ -36,7 +36,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="reqs">Conditions requirements that must be met</param>
         /// <param name="npcId">NPCs that drop the loot</param>
         /// <param name="itemId">Item ID to listen for</param>
-        private NpcDropCondition(ConditionReqs reqs, short npcId, int itemId) : base($"{CustomName}_{string.Join(",", npcId)}", reqs, [itemId]) => Listen(this, npcId);
+        private NpcDropCondition(ConditionReqs reqs, int npcId, int itemId) : base($"{CustomName}_{string.Join(",", npcId)}", reqs, [itemId]) => Listen(this, npcId);
 
         /// <summary>
         /// Creates a condition that listens for the NPC to drop any of the items
@@ -44,7 +44,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="reqs">Conditions requirements that must be met</param>
         /// <param name="npcId">NPCs that drop the loot</param>
         /// <param name="itemIds">Item IDs to listen for</param>
-        private NpcDropCondition(ConditionReqs reqs, short npcId, int[] itemIds) : base($"{CustomName}_{string.Join(",", npcId)}", reqs, itemIds) => Listen(this, npcId);
+        private NpcDropCondition(ConditionReqs reqs, int npcId, int[] itemIds) : base($"{CustomName}_{string.Join(",", npcId)}", reqs, itemIds) => Listen(this, npcId);
 
 
         /// <summary>
@@ -54,7 +54,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="npcId">NPC ID that drops the item</param>
         /// <param name="itemId">Item ID to listen for</param>
         /// <returns>NPC item drop achievement condition</returns>
-        public static CustomAchievementCondition Drop(ConditionReqs reqs, short npcId, int itemId) => new NpcDropCondition(reqs, npcId, itemId);
+        public static CustomAchievementCondition Drop(ConditionReqs reqs, int npcId, int itemId) => new NpcDropCondition(reqs, npcId, itemId);
 
         /// <summary>
         /// Helper to create a condition that listens for the NPC to drop any of the items
@@ -63,7 +63,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="npcId">NPC ID that drops the item</param>
         /// <param name="itemIds">Item IDs to listen for</param>
         /// <returns>NPC item drop achievement condition</returns>
-        public static CustomAchievementCondition DropAny(ConditionReqs reqs, short npcId, params int[] itemIds) => new NpcDropCondition(reqs, npcId, itemIds);
+        public static CustomAchievementCondition DropAny(ConditionReqs reqs, int npcId, params int[] itemIds) => new NpcDropCondition(reqs, npcId, itemIds);
 
         /// <summary>
         /// Helper to create a condition that listens for the NPC to drop all of the items
@@ -72,7 +72,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="npcId">NPC ID that drops the item</param>
         /// <param name="itemIds">Item IDs to listen for</param>
         /// <returns>NPC item drop achievement conditions</returns>
-        public static List<CustomAchievementCondition> DropAll(ConditionReqs reqs, short npcId, params int[] itemIds)
+        public static List<CustomAchievementCondition> DropAll(ConditionReqs reqs, int npcId, params int[] itemIds)
         {
             List<CustomAchievementCondition> conditions = [];
             foreach (var itemId in itemIds)
@@ -86,7 +86,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// <param name="player">Player that damaged the NPC</param>
         /// <param name="npcId">NPC ID that dropped the item</param>
         /// <param name="itemId">Item ID that was dropped</param>
-        private static void CustomAchievementsHelper_OnNpcDrop(Player player, short npcId, int itemId)
+        private static void CustomAchievementsHelper_OnNpcDrop(Player player, int npcId, int itemId)
         {
             if (!IsListeningForId(itemId, _listeners, out var conditions))
                 return;
@@ -106,7 +106,7 @@ namespace TerrariaAchievementLib.Achievements.Conditions
         /// Listen for events so the condition can be completed
         /// </summary>
         /// <param name="npcId">NPC ID that should drop the loot</param>
-        private static void Listen(NpcDropCondition condition, short npcId)
+        private static void Listen(NpcDropCondition condition, int npcId)
         {
             if (!_isHooked)
             {
